First clean-up.
