+++ light-client-core/src/main/scala/ch/epfl/ognjanovic/stevan/tendermint/verified/types/ValidatorSet.scala
@@ -91,11 +91,11 @@ object ValidatorSet {
         removeOne(h, first)
         interestingEquality(h, first, t)
         listSetRemoveHeadSameAsSubtraction(second)
-        removingContainment(h, first, second)
+        ListUtils.removingContainment(h, first, second)
         sumWithDifferenceIsEqual(removed, t)
 
       case Cons(_, t) =>
-        doesNotHaveHeadContainedInTail(first, second)
+        ListUtils.doesNotHaveHeadContainedInTail(first, second)
         sumWithDifferenceIsEqual(first, t)
     }
   }.ensuring(_ => sumVotingPower(first) + sumVotingPower(second -- first) == sumVotingPower(second))
@@ -106,7 +106,7 @@ object ValidatorSet {
     require(ListOps.noDuplicate(list) && list.contains(elem) && list.nonEmpty)
     list match {
       case Cons(_, Nil()) => ()
-      case Cons(_, tail) if !tail.contains(elem) => removingNonContained(tail, elem)
+      case Cons(_, tail) if !tail.contains(elem) => ListUtils.removingNonContained(tail, elem)
       case Cons(_, tail) => removeOne(elem, tail)
     }
   }.ensuring(_ =>
