+++ stainless.conf
@@ -1,4 +1,4 @@
-vc-cache = true
+vc-cache = false
 timeout = 300
 check-models = true
 print-ids = false
@@ -9,4 +9,4 @@ fail-invalid = false
 infer-measures = true
 check-measures = true
 type-checker = true
-//functions = [stepByStepVerification]
+//functions = [setIntersectionLemma]
